Made physical-part-detail into physical-part-detail-panel
