+++ frontend/web/src/app/components/physical-part/physical-part-detail-panel/physical-part-detail-panel.component.spec.ts
@@ -1,5 +1,11 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
+import { ReactiveFormsModule } from '@angular/forms';
+import { RouterTestingModule } from '@angular/router/testing';
+import { BsModalService } from 'ngx-bootstrap';
+import { PhysicalPartMockService } from '../../../services/physicalPart/physical-part-mock.service';
+import { PhysicalPartService } from '../../../services/physicalPart/physical-part.service';
+import { ErrorService } from '../../../services/error.service';
 import { PhysicalPartDetailPanelComponent } from './physical-part-detail-panel.component';
 
 describe('PhysicalPartDetailPanelComponent', () => {
@@ -9,6 +15,16 @@ describe('PhysicalPartDetailPanelComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [PhysicalPartDetailPanelComponent],
+      imports: [ReactiveFormsModule, RouterTestingModule],
+    });
+    TestBed.overrideComponent(PhysicalPartDetailPanelComponent, {
+      set: {
+        providers: [
+          { provide: PhysicalPartService, useClass: PhysicalPartMockService },
+          { provide: ErrorService, useClass: ErrorService },
+          BsModalService,
+        ],
+      },
     })
       .compileComponents();
   }));
@@ -19,7 +35,7 @@ describe('PhysicalPartDetailPanelComponent', () => {
     fixture.detectChanges();
   });
 
-  it('should PhysicalPartDetailPanelComponent', () => {
+  it('should create', () => {
     expect(component).toBeTruthy();
   });
 });
