--- frontend/web/src/app/components/physical-part/physical-part-detail/physical-part-detail.component.html
@@ -1,39 +1,0 @@
-<div class="container">
-
-  <div class="row">
-    <div *ngIf="loaded" >
-
-      <h2  *ngIf="!creating">Editing PhysicalPart</h2>
-      <h2  *ngIf="creating">Creating New PhysicalPart</h2>
-
-      <hr>
-
-      <div class="panel panel-default">
-        <div class="panel-body" style="padding-right: 30px">
-
-          <form *ngIf="requiredFieldsForm" [formGroup]="requiredFieldsForm">
-            <label>Name</label>
-            <div [ngClass]="{'has-error': !requiredFieldsForm.get('name').valid}">
-              <input class="form-control" formControlName="name" value="{{physicalPart.name}}">
-            </div>
-
-
-            <label>Email</label>
-            <div [ngClass]="{'has-error': !requiredFieldsForm.get('eMail').valid}">
-              <input class="form-control" formControlName="eMail" value="{{physicalPart.eMail}}">
-            </div>
-          </form>
-
-          <div class="pull-left" style="margin-top: 20px">
-            <button class="btn btn-default"
-                    style="min-width: 100px" (click)="cancel()">Cancel</button>
-          </div>
-          <div class="pull-right" style="margin-top: 20px">
-            <button [disabled]=" !requiredFieldsForm.valid" class="btn btn-primary"
-                    style="min-width: 100px" (click)="save()">Save</button>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
