+++ frontend/web/src/app/components/physical-part/physical-part-master-view/physical-part-master-view.component.html
@@ -5,12 +5,12 @@
   </div>
   <div class="row">
     <div class="col-md-7" id="list">
-      <app-physicalPart-list (selected)="physicalPartSelected($event)">
-      </app-physicalPart-list>
+      <app-physical-part-list (selected)="physicalPartSelected($event)">
+      </app-physical-part-list>
     </div>
     <div class="col-md-5" id="detail" style="margin-top: 54px">
-      <app-physicalPart-detail-panel [physicalPart]="selectedPhysicalPart">
-      </app-physicalPart-detail-panel>
+      <app-physical-part-detail-panel [physicalPart]="selectedPhysicalPart">
+      </app-physical-part-detail-panel>
     </div>
 
   </div>
